changing-wallpaper: Changes to Allure by Positron Dream
Adds credit to the image (Positron Dream)

Changes how to center the name. Sets margin to auto to horizontally
align the element. Sets position to absolute and top to 50% to
vertically align the top border. To vertically align, apply a
transform(0, -50%) to move upwards by half the size of the element